+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Compatibility:
 * Handle `Object#autoload` when autoload itself (#1616, @ssnickolay)
 * Skip upgraded default gems while loading RubyGems (#2075).
 * Verify that gem paths are correct before loading RubyGems (#2075).
+* Implement `rb_ivar_count`.
 
 Performance: